git-svn-id: https://scuba.svn.sourceforge.net/svnroot/scuba@72 b350d931-f925-4dad-8688-0e260e346352

+++ scuba_smartcards_j2se/src/net/sourceforge/scuba/smartcards/CardFileInputStream.java
@@ -146,7 +146,7 @@ public class CardFileInputStream extends InputStream
 			if (!Arrays.equals(fs.getSelectedPath(), path)) {
 				for (FileInfo fileInfo: path) { fs.selectFile(fileInfo.getFID()); }
 			}
-			byte[] data = fs.readBinary((short) offsetInFile, le);
+			byte[] data = fs.readBinary(offsetInFile, le);
 			System.arraycopy(data, 0, buffer, 0, data.length);
 			return data.length;
 		}
